Fix LX01 do not delete sounds if folder not found
It destroyed all my project data. :'D

+++ scripts/40_sounds.sh
@@ -1,21 +1,35 @@
 #!/bin/sh
 
-echo "[*] Mixing all sounds"
-
 BACK=$PWD
 USR_SHARE=$ROOTFS/usr/share
-cd $USR_SHARE/sound-vendor
 
-for FOLDER in *; do
+if [ -d "$USR_SHARE/sound-vendor" ]; then
+  echo "[*] Mixing all sound-vendors"
+
+  SOUND_FOLDER=$USR_SHARE/sound-vendor
+  cd $SOUND_FOLDER
+
+  for FOLDER in *; do
+    mkdir -p original
+    mv -vf $FOLDER/* original/
+    rm -rf $FOLDER
+    ln -sf original $FOLDER
+  done
+elif [ -d "$USR_SHARE/sound" ]; then
+  echo "[*] Moving sound content"
+
+  SOUND_FOLDER=$USR_SHARE/sound
+  cd $SOUND_FOLDER
+
   mkdir -p original
-  mv -vf $FOLDER/* original/
-  rm -rf $FOLDER
-  ln -sf original $FOLDER
-done
+  mv -vf * original/
+else
+  echo "[!] Sound folder not detected! Skipping."
+  exit
+fi
 
 echo "[*] Remove original chinese voices"
 cd $BACK
-SOUND_FOLDER=$ROOTFS/usr/share/sound-vendor
 
 for NAME in aux_on aux_off aux_stop_tip \
   bluetooth_already_connected bluetooth_connect bluetooth_disconnect bluetooth_failure \
@@ -25,17 +39,28 @@ for NAME in aux_on aux_off aux_stop_tip \
   reset_wait reset service_timeout setup_failure tts_vendor_demo unknown_action unknown_domain \
   unknown_service upgrade_later upgrade_now wakeup_ei_01 wakeup_ei_02 wakeup_mitv wakeup_wozai_01 \
   wakeup_zai_01 wakeup_zai_02 weak_network welcome wifi_disconnect; do
-  rm -vf $USR_SHARE/sound-vendor/original/$NAME.*
+  rm -vf $SOUND_FOLDER/original/$NAME.*
 done
 
-echo "[*] Creating sound directory and linking factory sounds"
 cd $BACK
+
+if [ -d "$USR_SHARE/sound-vendor" ]; then
+  SOUND_FOLDER=sound-vendor
+
+  echo "[*] Creating sound directory"
+  cd $USR_SHARE
+  rm -f sound
+  mkdir -p sound
+  cd $BACK
+else
+  SOUND_FOLDER=sound
+fi
+
+echo "[*] Linking factory sounds"
 cd $USR_SHARE
-rm -f sound
-mkdir -p sound
-for FILE in sound-vendor/original/*; do
+for FILE in $SOUND_FOLDER/original/*; do
   NAME=$(basename $FILE)
-  ln -sf /usr/share/sound-vendor/original/$NAME sound/$NAME
+  ln -sf /usr/share/$SOUND_FOLDER/original/$NAME sound/$NAME
 done
 
 cd $BACK
